adjust

+++ src/main/java/codedriver/module/cmdb/dao/mapper/cientity/AttrEntityMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright(c) 2021 TechSure Co., Ltd. All Rights Reserved.
+ * Copyright(c) 2022 TechSure Co., Ltd. All Rights Reserved.
  * 本内容仅限于深圳市赞悦科技有限公司内部传阅，禁止外泄以及用于其他的商业项目。
  */
 
@@ -16,7 +16,7 @@ import java.util.List;
 public interface AttrEntityMapper {
     List<AttrEntityVo> getAttrEntityByFromCiIdAndAttrId(AttrEntityVo attrEntityVo);
 
-    List<AttrEntityVo> getAttrEntityByFromCiEntityIdAndAttrId(@Param("fromCiEntityId") Long fromCiEntityId, @Param("attrId") Long attrId, @Param("limit") Integer limit);
+    List<AttrEntityVo> getAttrEntityByFromCiEntityIdAndAttrId(@Param("fromCiEntityId") Long fromCiEntityId, @Param("attrId") Long attrId, @Param("limit") Long limit);
 
     void updateAttrEntityFromIndex(AttrEntityVo attrEntityVo);
