+++ src/main/java/codedriver/module/cmdb/dao/mapper/cientity/RelEntityMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright(c) 2021 TechSure Co., Ltd. All Rights Reserved.
+ * Copyright(c) 2022 TechSure Co., Ltd. All Rights Reserved.
  * 本内容仅限于深圳市赞悦科技有限公司内部传阅，禁止外泄以及用于其他的商业项目。
  */
 
@@ -41,12 +41,12 @@ public interface RelEntityMapper {
                                                       @Param("relIdList") List<Long> relIdList);
 
     List<RelEntityVo> getRelEntityByFromCiEntityIdAndRelId(@Param("fromCiEntityId") Long fromCiEntityId,
-                                                           @Param("relId") Long relId, @Param("limit") Integer limit);
+                                                           @Param("relId") Long relId, @Param("limit") Long limit);
 
     RelEntityVo getRelEntityByFromCiEntityIdAndToCiEntityIdAndRelId(@Param("fromCiEntityId") Long fromCiEntityId, @Param("toCiEntityId") Long toCiEntityId, @Param("relId") Long relId);
 
     List<RelEntityVo> getRelEntityByToCiEntityIdAndRelId(@Param("toCiEntityId") Long toCiEntityId,
-                                                         @Param("relId") Long relId, @Param("limit") Integer limit);
+                                                         @Param("relId") Long relId, @Param("limit") Long limit);
 
     List<Long> getFromToCiEntityIdByCiEntityIdList(@Param("idList") List<Long> idList);
