+++ src/main/java/codedriver/module/cmdb/publicapi/GetCiEntityAttrEntityApi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright(c) 2021 TechSure Co., Ltd. All Rights Reserved.
+ * Copyright(c) 2022 TechSure Co., Ltd. All Rights Reserved.
  * 本内容仅限于深圳市赞悦科技有限公司内部传阅，禁止外泄以及用于其他的商业项目。
  */
 
@@ -120,7 +120,7 @@ public class GetCiEntityAttrEntityApi extends PublicApiComponentBase {
         CiEntityVo returnCiEntityVo = new CiEntityBuilder.Builder(ciEntityVo, resultList, ciVo, attrList, null).build().getCiEntity();
         if (returnCiEntityVo != null) {
             //拼接引用属性数据
-            Integer attrEntityLimit = null;
+            Long attrEntityLimit = null;
             if (CollectionUtils.isNotEmpty(attrList)) {
                 for (AttrVo attrVo : attrList) {
                     if (attrVo.getTargetCiId() != null) {
